move stagehand out of index.ts
