+++ lib/stagehand/context.ts
